Add files via upload
In this chapter, we develop several architectures based on deep neural networks (DNN). Through discussing their advantages and drawbacks, we try to develop more appropriate models to meet the demands of the British Hypertension Society and the Association for the Advancement of Medical Instrumentation standards.  The building blocks of our models are convolutional neural networks (CNN) and long short-term memories (LSTM). We evaluate our models on 200 subjects.
The following models for BP estimation are discussed here: 
1.	LSTM Model
2.	PCA-LSTM Model
3.	CNN Model
4.	CNN-LSTM Model
To access the processed data used here, refer to the following “Google Drive” links:
https://drive.google.com/file/d/18vgqywmqlArXRX5M7l_WJSd5WH8dM1xe/view?usp=sharing
For more questions, contact me by email: JEsmaelpoor@gmail.com